Switch to using a context-free grammar for Tannot
Now Tannot carries an optional context-free grammar (CFG). This can be
used to specify what shape the values have in evaluate-based
proofs. It is simpler than using datatype definitions directly.

However, this draft solution is not satisfactory:
 - the definition of how to traverse values is a bit too clunky
 - the type system checks CakeML code *generated* from the CFG
 - but the generated CakeML code is *not* what evaluate runs (because
   then we'd have to consume clock ticks, which we can't afford since
   consuming clock ticks would mean that the compiler can't throw away
   these type annotations)
 - the changes to the functions are only in the HOL files, not in the
   Lem files (we can edit the Lem if these changes are to stay)

I suspect we shouldn't add this feature to the source semantics. I
will discuss this with Oskar and suggest another way forward.