Draft a variant of Tannot with exp as annot
Oskar requested that I commit a (failed?) experiment. Here it is.

The idea is that type annotations carry an extra expression
that must evaluate to unit. The expression's effect is ignored and
the compiler is to not produce code for the new expression
annotation. However, any type errors produced by the annotation
expression are propagated.

I worry that the two remaining cheats in evaluateProps are
not false.