Move a concrete env with sem_env
This is needed for the 'abstract compiler' which is currently the
goal of the work. It will sit in the semantics and represent the
actual CakeML incremental compiler. We need to provide it with an
environment argument, which the user can pick, but the problem is
that the environment as represented by the compiler is quite
different to the compiler as represented by the semantics. To try
to solve this we move the two in parallel, each Env object having
both the semantic environment and an environment value.